rename applications
