+++ jpa/multitenancy/multitenancy-db/pom.xml
@@ -213,7 +213,7 @@
                 <configuration>
                     <java>
                         <googleJavaFormat>
-                            <version>1.25.0</version>
+                            <version>1.25.2</version>
                             <style>AOSP</style>
                         </googleJavaFormat>
                     </java>
