+++ jpa/multitenancy/multitenancy-db/src/main/java/com/example/multitenancy/db/MultiTenancyDBApplication.java
@@ -7,9 +7,9 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 
 @SpringBootApplication
 @EnableConfigurationProperties({ApplicationProperties.class})
-public class Application {
+public class MultiTenancyDBApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
+        SpringApplication.run(MultiTenancyDBApplication.class, args);
     }
 }
