+++ jpa/multitenancy/partition/src/test/java/com/example/multitenancy/partition/TestMultiTenancyPartitionApplication.java
@@ -3,9 +3,11 @@ package com.example.multitenancy.partition;
 import com.example.multitenancy.partition.common.ContainersConfig;
 import org.springframework.boot.SpringApplication;
 
-public class TestApplication {
+public class TestMultiTenancyPartitionApplication {
 
     public static void main(String[] args) {
-        SpringApplication.from(Application::main).with(ContainersConfig.class).run(args);
+        SpringApplication.from(MultiTenancyPartitionApplication::main)
+                .with(ContainersConfig.class)
+                .run(args);
     }
 }
