+++ jpa/multitenancy/schema/src/main/java/com/example/multitenancy/schema/MultitenancySchemaApplication.java
@@ -1,15 +1,15 @@
-package com.example.multitenancy.partition;
+package com.example.multitenancy.schema;
 
-import com.example.multitenancy.partition.config.ApplicationProperties;
+import com.example.multitenancy.schema.config.ApplicationProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
 @SpringBootApplication
 @EnableConfigurationProperties({ApplicationProperties.class})
-public class Application {
+public class MultitenancySchemaApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
+        SpringApplication.run(MultitenancySchemaApplication.class, args);
     }
 }
