+++ jpa/multitenancy/schema/src/test/java/com/example/multitenancy/schema/TestMultitenancySchemaApplication.java
@@ -3,9 +3,11 @@ package com.example.multitenancy.schema;
 import com.example.multitenancy.schema.config.TestContainersConfiguration;
 import org.springframework.boot.SpringApplication;
 
-public class TestApplication {
+public class TestMultitenancySchemaApplication {
 
     public static void main(String[] args) {
-        SpringApplication.from(Application::main).with(TestContainersConfiguration.class).run(args);
+        SpringApplication.from(MultitenancySchemaApplication::main)
+                .with(TestContainersConfiguration.class)
+                .run(args);
     }
 }
